separate cohere functions to another module
